Prep for release v0.0.2
Add test for Postgres 17 and increment the version to v0.0.2.

Add a GitHub workflow to auto-publish on PGXN on release. To use it, add
your PGXN username and password to to your [GitHub Actions secrets].
Create these environment secrets:

*   PGXN_USERNAME
*   PGXN_PASSWORD

It should then publish when you push the tag `v0.0.2`.

While at it, ignore generated Bitcode (`.bc`) and darwin (`.dylb`)
files.

[GitHub Actions secrets]: https://github.com/adjust/pg-base62/settings/secrets/actions